Fix issue with mkv file player and video quality adaptation, causing video player to be replaced by camera by accident.

+++ src/voip/private.h
@@ -102,7 +102,7 @@ bool_t ms_media_stream_sessions_secured(const MSMediaStreamSessions *sessions,Me
 
 /* The handler of tmmbr for MediaStream objects.*/
 void media_stream_tmmbr_received(const OrtpEventData *evd, void *user_pointer);
-void media_stream_process_tmmbr(MediaStream *ms, int tmmbr_mxtbr);
+void media_stream_process_tmmbr(MediaStream *ms, uint64_t tmmbr_mxtbr);
 
 void media_stream_add_tmmbr_handler(MediaStream *stream, void (*on_tmmbr_received)(const OrtpEventData *evd, void *), void *user_data);
